Move restore tests to integration tests
